Add ScheduleEntry service and adjust related models
Implemented a new ScheduleEntry service that provides basic CRUD operations for schedules. This includes adding, getting, removing and updating schedules. Additionally, the ScheduleEntryDTO model was adjusted to replace the SPLCorrelation object with SPLCorrelationId. Also, error messages were extended to better handle situations when a schedule or a stop point line correlation not found.