Implement API versioning and add StopPointController
The API versioning has been configured with the introduction of 'ApiVersion' attribute and 'ApiExplorerSettings' for controllers. This adds hierarchical grouping in the Swagger UI documentation. The data is received from the query in RemoveSchedule method. A new controller, StopPointController, has been added which includes basic CRUD operations. Furthermore, the OpenApi description of Swagger has been customized.